refactor: create InfoBox component
